Create prime-bitpacked.pas
Using the "packed" keyword does not put 8 booleans in one byte. For that you need to use "bitpacked" or some compiler directive. But then the compiler won't accept dynamic arrays. Here is an implementation using a pointer to bitpacked array. I've tested with actual bit-packed array, it was about the same speed for 1 million, but for 100 million, the bit-packed version was much faster, I guess due to reduced memory cache usage.